Move lang.observe and observer implementations to their own files
