+++ tests/unit/all.ts
@@ -11,8 +11,8 @@ import './lang';
 import './math';
 import './Map';
 import './number';
-import './observers/ObjectObserver';
 import './object';
+import './observe';
 import './on/all';
 import './Promise';
 import './Registry';
