+++ tests/unit/lang.ts
@@ -1,9 +1,6 @@
 import registerSuite = require('intern!object');
 import assert = require('intern/chai!assert');
-import has from 'src/has';
 import * as lang from 'src/lang';
-import { PropertyEvent } from 'src/observers/interfaces';
-import { Es5Observer, Es7Observer } from 'src/observers/ObjectObserver';
 
 registerSuite({
 	name: 'lang functions',
@@ -165,46 +162,6 @@ registerSuite({
 		assert.isFalse(Object.getOwnPropertyDescriptor(copyOfObject, 'c').configurable);
 	},
 
-	'.observe() when nextTurn is true': function () {
-		if (!has('object-observe')) {
-			this.skip('Native Object.observe support is required for this test.');
-		}
-
-		const observer = lang.observe({
-			target: {},
-			listener: function () {},
-			nextTurn: true
-		});
-		assert.isTrue(observer instanceof Es7Observer);
-	},
-
-	'.observe() when nextTurn is false': function () {
-		if (!has('object-observe')) {
-			this.skip('Native Object.observe support is required for this test.');
-		}
-
-		const observer = lang.observe({
-			target: {},
-			listener: function () {},
-			nextTurn: false
-		});
-		assert.isTrue(observer instanceof Es5Observer);
-	},
-
-	'.observe() onlyReportObserved': function () {
-		if (!has('object-observe')) {
-			this.skip('Native Object.observe support is required for this test.');
-		}
-
-		const observer = lang.observe({
-			target: {},
-			listener: function () {},
-			nextTurn: true,
-			onlyReportObserved: true
-		});
-		assert.isTrue(observer.onlyReportObserved, 'onlyReportObserved should be passed to the observer instance.');
-	},
-
 	'.getPropertyDescriptor()': function () {
 		const object1 = {
 			get foo() { return 'bar'; }
@@ -251,6 +208,9 @@ registerSuite({
 		assert.isFalse(lang.isIdentical(3, NaN));
 		assert.isFalse(lang.isIdentical(3, '3'));
 		assert.isTrue(lang.isIdentical(Infinity, Infinity));
+		assert.isFalse(lang.isIdentical(-0, 0));
+		assert.isTrue(lang.isIdentical(0, 0));
+		assert.isTrue(lang.isIdentical(-0, -0));
 	},
 
 	'.lateBind() context': function () {
