+++ tsconfig.json
@@ -37,8 +37,7 @@
 		"./src/math.ts",
 		"./src/number.ts",
 		"./src/object.ts",
-		"./src/observers/ObjectObserver.ts",
-		"./src/observers/interfaces.d.ts",
+		"./src/observe.ts",
 		"./src/on.ts",
 		"./src/queue.ts",
 		"./src/request.ts",
@@ -108,7 +107,7 @@
 		"./tests/unit/module.d.ts",
 		"./tests/unit/number.ts",
 		"./tests/unit/object.ts",
-		"./tests/unit/observers/ObjectObserver.ts",
+		"./tests/unit/observe.ts",
 		"./tests/unit/on/all.ts",
 		"./tests/unit/on/browserOnly.ts",
 		"./tests/unit/on/common.ts",
